+++ cmd/agent/app/flags_test.go
@@ -24,35 +24,39 @@ import (
 	"flag"
 	"testing"
 
+	"github.com/spf13/cobra"
+	"github.com/spf13/viper"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
 func TestBingFlags(t *testing.T) {
-	cfg := NewBuilder()
-	flags := flag.NewFlagSet("test", flag.ContinueOnError)
-	cfg.Bind(flags)
-	err := flags.Parse([]string{
-		"-collector.host-port=1.2.3.4:555,1.2.3.4:666",
-		"-discovery.min-peers=42",
-		"-http-server.host-port=:8080",
-		"-processor.jaeger-binary.server-host-port=:1111",
-		"-processor.jaeger-binary.server-max-packet-size=4242",
-		"-processor.jaeger-binary.server-queue-size=42",
-		"-processor.jaeger-binary.workers=42",
-	})
+	v := viper.New()
+	b := &Builder{}
+	command := cobra.Command{}
+	flags := &flag.FlagSet{}
+	AddFlags(flags)
+	command.PersistentFlags().AddGoFlagSet(flags)
+	v.BindPFlags(command.PersistentFlags())
 
-	assert.NoError(t, err)
-	assert.Equal(t, 3, len(cfg.Processors))
-	assert.Equal(t, []string{"1.2.3.4:555", "1.2.3.4:666"}, cfg.CollectorHostPorts)
-	assert.Equal(t, 42, cfg.DiscoveryMinPeers)
-	assert.Equal(t, ":8080", cfg.HTTPServer.HostPort)
-	assert.Equal(t, ":1111", cfg.Processors[2].Server.HostPort)
-	assert.Equal(t, 4242, cfg.Processors[2].Server.MaxPacketSize)
-	assert.Equal(t, 42, cfg.Processors[2].Server.QueueSize)
-	assert.Equal(t, 42, cfg.Processors[2].Workers)
-
-	err = flags.Parse([]string{
-		"-collector.host-port=1.2.3.4:555,1.2.3.4:666",
+	err := command.ParseFlags([]string{
+		"--collector.host-port=1.2.3.4:555,1.2.3.4:666",
+		"--discovery.min-peers=42",
+		"--http-server.host-port=:8080",
+		"--processor.jaeger-binary.server-host-port=:1111",
+		"--processor.jaeger-binary.server-max-packet-size=4242",
+		"--processor.jaeger-binary.server-queue-size=42",
+		"--processor.jaeger-binary.workers=42",
 	})
-	assert.NotNil(t, err)
+	require.NoError(t, err)
+
+	b.InitFromViper(v)
+	assert.Equal(t, 3, len(b.Processors))
+	assert.Equal(t, []string{"1.2.3.4:555", "1.2.3.4:666"}, b.CollectorHostPorts)
+	assert.Equal(t, 42, b.DiscoveryMinPeers)
+	assert.Equal(t, ":8080", b.HTTPServer.HostPort)
+	assert.Equal(t, ":1111", b.Processors[2].Server.HostPort)
+	assert.Equal(t, 4242, b.Processors[2].Server.MaxPacketSize)
+	assert.Equal(t, 42, b.Processors[2].Server.QueueSize)
+	assert.Equal(t, 42, b.Processors[2].Workers)
 }