+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Changes by Version
 ==================
 
+0.5.2 (2017-07-20)
+------------------
+
+- Use official Cassandra 3.11 base image [#278](https://github.com/uber/jaeger/pull/278)
+- Support configurable metrics backend in the agent [#275](https://github.com/uber/jaeger/pull/275)
+
+
 0.5.1 (2017-07-03)
 ------------------